Exported and renamed a few packages in the netatmo binding. See openhab/static-code-analysis#100
